Interactive map working
Option2 for Please make the following changes 1) The interactive map shall have more space. Adjust the layout. The calculation output shall be below the manual input and the map. 2) Explanation will follow next. 3) Result and Explanation for time to degree calculation shall not be same as degree to time. For time to longitude the time shall come first and vice versa.4) Selected longitude should change based on manual input or based on sliding input.5) There shall be a cap of +/-12 for the time input. Similarly +/- 180 degree for the longitudinal input and the slider