Improved timezone converter
yes but need not be there in the explanation. Also this has to be elaborated Decimal degrees → D:M:S = 121° 0' 0.000" E. Can we put the result under the manual input, explanation can be given after the map. Somewhere we have to incorporate both in input and result that the time is UTCoffseet UTC+/-